+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Fixed bug where child account's DCEPrincipal role trusted itself rather than the master account
 - Add GetUsageByPrincipal
 - Fix default `budget_notification_from_email` TF var (See #143)
-
+- Add GetUsage with filter options: accountId, principalId, startDate, limit
 
 ## v0.23.0